masking-service/requirements.txt
fastapi
uvicorn
requests
python-multipart
easyocr
numpy
opencv-python-headless
Pillow
httpx
--extra-index-url https://download.pytorch.org/whl/cpu
torch
boto3